+++ NEWS.md
@@ -4,6 +4,7 @@
 - Rename 'codelists' to 'cells' in px-object and Excel workbooks. (#256)
 - Deprecate 'metamake', 'pxmake' and 'make_template'. (#198)
 - Change name of all modifying functions by adding the prefix 'px_'. (#254)
+- Rename 'acrosscell' to 'acrosscells'. (#271)
 
 ## New features
 - Add modifying functions: `map()`, `baseperiod()`, `domain()`, `elimination()`,